Rename health_check_celery plugins
Celery 3 and 4 are the currently used versions,
so rename health_check_celery3 to health_check_celery.

Rename health_check_celery to health_check_celery2.

Add change notices to README.

Upgrading should be very easy for existing users,
it will simply require renaming the imported
celery package in existing applications'
INSTALLED_APPS.
