+++ health_check_celery/plugin_health_check.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import logging
 from datetime import datetime, timedelta
 from time import sleep
 
@@ -10,21 +11,40 @@ from health_check.backends.base import (
 from health_check.plugins import plugin_dir
 from health_check_celery.tasks import add
 
+logger = logging.getLogger(__name__)
+
 
 class CeleryHealthCheck(BaseHealthCheckBackend):
 
     def check_status(self):
         timeout = getattr(settings, 'HEALTHCHECK_CELERY_TIMEOUT', 3)
-        expires = datetime.now() + timedelta(seconds=timeout)
+
         try:
-            result = add.apply_async(args=[4, 4], expires=expires, connect_timeout=timeout)
+            result = add.apply_async(
+                args=[4, 4],
+                expires=datetime.now() + timedelta(seconds=timeout)
+            )
             now = datetime.now()
             while (now + timedelta(seconds=3)) > datetime.now():
-                if result.result == 8:
+                print("            checking....")
+                if result.ready():
+                    try:
+                        result.forget()
+                    except NotImplementedError:
+                        pass
                     return True
                 sleep(0.5)
         except IOError:
-            pass
+            logger.exception("IOError")
+            raise ServiceUnavailable("IOError")
+        except:
+            logger.exception("Unknown Error")
+            raise ServiceUnavailable("Unknown error")
+
+        logger.error(
+            u'Celery task did not complete successfully. '
+            u'Verify celery is running'
+        )
         raise ServiceUnavailable("Unknown error")
 
 plugin_dir.register(CeleryHealthCheck)
