+++ health_check_celery/tasks.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
-from celery.task import task
+from __future__ import absolute_import
+
+from celery import shared_task
 
 
-@task
+@shared_task(ignore_result=False)
 def add(x, y):
     return x + y
