+++ health_check_celery2/tasks.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+from celery.task import task
+
+
+@task
+def add(x, y):
+    return x + y
