--- health_check_celery3/tasks.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import
-
-from celery import shared_task
-
-
-@shared_task(ignore_result=False)
-def add(x, y):
-    return x + y
