+++ tests/testapp/settings.py
@@ -23,7 +23,7 @@ INSTALLED_APPS = (
     'health_check_db',
     'health_check_storage',
     'health_check_celery',
-    'health_check_celery3',
+    'health_check_celery2',
     'tests',
 )
